requirements.txt
selenium==3.141.0
beautifulsoup4==4.10.0
bs4==0.0.1
python-dotenv==0.19.2
pymongo==4.0.1
psycopg2==2.9.3
pypika==0.48.8